Restructure splitting many methods into separate files
