+++ Sources/JNI/JNIFields.swift
@@ -23,3 +23,80 @@ public extension JNI {
         return try T.fromField(fieldID!, on: javaObject)
     }
 }
+
+
+public extension JNI {
+    // MARK: Fields
+
+    public func GetBooleanField(of javaObject: JavaObject, id: JavaFieldID) throws -> JavaBoolean {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetBooleanField(_env, javaObject, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetIntField(of javaObject: JavaObject, id: JavaFieldID) throws -> JavaInt {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetIntField(_env, javaObject, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetFloatField(of javaObject: JavaObject, id: JavaFieldID) throws -> JavaFloat {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetFloatField(_env, javaObject, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetDoubleField(of javaObject: JavaObject, id: JavaFieldID) throws -> JavaDouble {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetDoubleField(_env, javaObject, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetObjectField(of javaObject: JavaObject, id: JavaFieldID) throws -> JavaObject {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetObjectField(_env, javaObject, id)
+        try checkAndThrowOnJNIError()
+        return result!
+    }
+
+    // MARK: Static Fields
+
+    public func GetStaticBooleanField(of javaClass: JavaClass, id: JavaFieldID) throws -> JavaBoolean {
+        let _env = self._env
+        let result =  _env.pointee.pointee.GetStaticBooleanField(_env, javaClass, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetStaticIntField(of javaClass: JavaClass, id: JavaFieldID) throws -> JavaInt {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetStaticIntField(_env, javaClass, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetStaticFloatField(of javaClass: JavaClass, id: JavaFieldID) throws -> JavaFloat {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetStaticFloatField(_env, javaClass, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetStaticDoubleField(of javaClass: JavaClass, id: JavaFieldID) throws -> JavaDouble {
+        let _env = self._env
+        let result = _env.pointee.pointee.GetStaticDoubleField(_env, javaClass, id)
+        try checkAndThrowOnJNIError()
+        return result
+    }
+
+    public func GetStaticObjectField(of javaClass: JavaClass, id: JavaFieldID) throws -> JavaObject {
+        let _env = self._env
+        guard let result = _env.pointee.pointee.GetStaticObjectField(_env, javaClass, id) else { throw JNIError() }
+        try checkAndThrowOnJNIError()
+        return result
+    }
+}
